V1.0 Working release - Tested, examples, and keywords
Created a simple example for the library that tested a majority of the functionality.  Further testing will continue to occur.

An example was created and put into the correct folder to allow it to show up in the arduino IDE.

Keyword file was created to allow syntax highlighting to occur on the libraries methods, constants, etc.