Updated usb_inst.cpp to add Serial fix
Issues with arduino libraries making calls to serial.  Even though there is no serial interface the teensyduino core files allows a "dummy" serial interface to be created so that these libraries still work and don't cause compilation errors.  This now works as intended and doesn't throw errors when standard libraries that print to a serial port for debugging are used.